+++ CHANGELOG.md
@@ -4,8 +4,33 @@
 
 ---
 
+## July 17, 2021
+
 ### Added
 
+- Add superuser token support
+
+### Changed
+
+- Refactor internal packages
+
+### Fixed
+
+## May 31, 2021
+
+### Added
+
+- Add chat topic in acl logics
+
+### Changed
+
+### Fixed
+
+## May 29, 2021
+
+### Added
+
+- Add location sharing topics acl logic.
 - Ignore tokens which contain `is_superuser` flag in their claims.
 - Tracing with jeager. jeager configuration is added